bundles: generate v2 stream clone bundles (Bug 1607915) r=glob,mhentges
This commit updates the clonebundle generation script to create v2 stream
clone bundles instead of a v1 bundle. The bundle generation command is
updated from `hg debugcreatestreamclonebundle` to `hg bundle` with arguments
for the bundle type, bringing it in line with other bundles. The bundlespec
is updated to reflect the format of new bundles as they would be created on
hgssh.

In the event this change causes unforeseen bustages in CI or elsewhere,
we can safely roll it back and revert clonebundle manifest files to the
previous day's version.

Depends on D59236

Differential Revision: https://phabricator.services.mozilla.com/D59237

--HG--
extra : moz-landing-system : lando