ansible: replace vct_dir stat check with "is_testing" variable (Bug 1603594) r=sheehan
It was previously implicit that, if "/vct" existed, we were running in a test environment.
By using a dedicated variable to represent this fact ("is_testing"), we can simplify the logic here.

Removes "ensure supervisor service is started" task from the "supervisor" role because
"supervisor" is only needed in the testing environment, where systemd doesn't exist.

Differential Revision: https://phabricator.services.mozilla.com/D59780

--HG--
extra : moz-landing-system : lando