commitparser: allow periods in reviewer names (Bug 1607521) r=glob,sheehan
This change updates the regular expression used to parse reviewer name to
align with moz-phab and Lando.

Also updated regex to allow for a single character reviewer name, as well as
added tests to cover various combinations of letters and periods.

Note that this does not currently support parsing names that end in a period
(whether that is including the period or stripping it).

Differential Revision: https://phabricator.services.mozilla.com/D58979

--HG--
extra : moz-landing-system : lando