ansible: stop using `with_items` to install packages via `yum` r=mhentges
When creating the `docker` role, Mitch noticed that using `with_items`
on newer Ansible versions caused warnings and recommended passing a list
of items as the `name` argument directly. We assume this optimizes `yum`,
which can better calculate the set of required dependencies when all
packages are specified up front.

Differential Revision: https://phabricator.services.mozilla.com/D60433

--HG--
extra : moz-landing-system : lando