vcsreplicator/bootstrap: use custom function to process extra bootstrap messages (Bug 1610635, Bug 1541166) r=mhentges
Between Python 2 and 3, the `map` builtin function has some slightly
modified behaviour that breaks the hgweb bootstrap processing of
extra messages, namely the parallelization of mapping in Python 3.

This commit switches from a fancy combination of `zip` and `map` to
create the future to instead use a simple sequential function to
apply the list of parameters to the `handle_message_main` function
in the correct order.

Differential Revision: https://phabricator.services.mozilla.com/D60601

--HG--
extra : moz-landing-system : lando