vcsreplicator: use stream clones in bootstrapping procedure (Bug 1610637) r=mhentges
At the moment the bootstrap procedure relies entirely on `hg pull`
to transfer repository data during the bootstrap procedure. This is
because bootstrap simply calls the same function from within it's
code that the replication system would call when running
`hg replicatesync` to force synchronize a repo on all hgweb mirrors.
This implementation detail causes bootstrapping to take much longer
than necessary.

This commit replaces the call to create an empty repository and then
`hg pull` the data into it, with a call to `hg clone --stream` that
falls back to an empty repo on failure. In production this should result
in the repo being seeded from a clonebundle using the Mozilla CDN,
which will dramatically speed up the bootstrap process.

We add a fallback to regular empty repository + `hg pull` due to the
test environment not having a clonebundle available for download. Since
there is no clonebundle to be used, the stream clone attempts to directly
access the underlying revlogs and other files in `.hg` directory of a
given repository, causing a permissions error. This should be avoided
in production, and in the worst case, we will fall back to the current
slow behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D60606

--HG--
extra : moz-landing-system : lando