hgmo: filter region specifier by referencing instance data object (Bug 1587958) r=mhentges,smacleod
Previously we referenced the instance data object to determine which
region the request originated from, and assumed AWS was the cloud
we were operating in. Now that we are moving into GCE as well, we
need to differentiate between the different clouds to correctly
serve bundles from those regions.

This commit adds a `cloud_region_specifier` function which takes a
parsed instance data dict and returns a byte-string that can be used
to filter clonebundles manifests to serve bundles from the same cloud
region as the originating request.

Differential Revision: https://phabricator.services.mozilla.com/D60704

--HG--
extra : moz-landing-system : lando