vcsreplicator/bootstrap: account for possible whitespace issues with Kafka host list r=zeid
The `replicationproducer.hosts` config option for vcsreplicator
producers and consumers is intended to be a comma separated list
of `host:port` string combinations. Our implementation uses the
`', '` string as the separator, and we do a `.split` of the raw
string on that substring to create the list of hosts.

As mentioned on Phabricator, this commit is a follow up to use
a slightly more resilient string parsing technique. With this
change we will avoid an unexpected breakage when someone removes
the trailing whitespace without understanding the implications.

Differential Revision: https://phabricator.services.mozilla.com/D60765

--HG--
extra : moz-landing-system : lando