configwizard: avoid error messages and delete Py3 bytecode caches in `test-legacy-version.t` r=mhentges
`test-legacy-version.t` is on the Python 3 failing list as `.pyc`
files are not created adjacent to their source under Py3. Thus our
call to remove the `.pyc` file in the same directory causes test
failures on Python 3, since those files don't exist.

This commit adds a `-f` flag to `rm`, which causes the removal to
fail silently. We also remove the `__pycache__` directory silently,
enforcing deterministic behaviour with regard to changes in the
`fakeversion.py` extension.

This pattern is mostly copied from `test-commandserver.t`[1] in upstream
Mercurial.

[1] https://www.mercurial-scm.org/repo/hg-committed/file/ae596fac8ba0/tests/test-commandserver.t#l260

Differential Revision: https://phabricator.services.mozilla.com/D60766

--HG--
extra : moz-landing-system : lando