robustcheckout: remove `supported_hg` check and mark as hg 5.3 compatible (Bug 1610856) r=zeid
The `supported_hg` check was added to robustcheckout to ensure any
consumers using the `sparse` extension features were on a new enough
version of Mercurial. Since then, all known clients in CI are running
versions of Mercurial that include this extension, this it has served
it's purpose.

On top of this, the check causes upgrades of Mercurial to fail quite
frequently as the version of `robustcheckout` used by the client also
needs to be updated to the latest. This means updating to a newer Mercurial
version will fail when the extension itself might be totally compatible,
but the newly released version is missing from the `testedwith` string.

Differential Revision: https://phabricator.services.mozilla.com/D64099

--HG--
extra : moz-landing-system : lando