hgtemplates: add Open Graph Meta tags to `gitweb_mozilla` (Bug 1614815) r=mhentges
This commit adds a useful set of Open Graph meta tags to each
of the rendered template pages of the `gitweb_mozilla` template
set. This change will cause sharing links to hgweb on Matrix to
render a more useful summary of the page content, as opposed to
the current output which mostly includes the content of the page
ribbon.

The common set of tags are added to the `header.tmpl` template,
which is rendered at the top of each page. Each page that can be
rendered by a Mercurial webcommand is also given a unique set of
headers which describe the page.

We add these new changes as a patch to `hgtemplates/.patches` so
the changes are taken into account when upgrading Mercurial.

Differential Revision: https://phabricator.services.mozilla.com/D65353

--HG--
extra : moz-landing-system : lando