ansible/test-kafkabroker: simplify `start-kafka` script (Bug 1600827) r=mhentges
The `start-kafka` test script, which is used to wait on Zookeeper
to be fully up and ready before starting Kafka on a host, is
littered with race conditions. The script waits on several indicators
that, after some research, seem to be mostly arbitrary and unnecessary
to wait on before starting Kafka. This commit simplifies the script
to instead call the `stat` command on each known Zookeeper endpoint
and waits until the response is not empty, which is the key metric
to assert a Zookeeper endpoint is ready to accept connections from
Kafka. We also add a check to wait on the `/kafka-servers` test file
to be populated before attempting to use it when starting Kafka.

Depends on D66159

Differential Revision: https://phabricator.services.mozilla.com/D66160

--HG--
extra : moz-landing-system : lando