testing: add a docker-compose file for standing up an hgmo test cluster (Bug 1600827) r=mhentges
This commit adds a `docker-compose` file to stand up a test
cluster for hg.mozilla.org. We define 5 services representing
the hgssh push server, the two hgweb servers, and
containers for Pulse and LDAP. Containers are started from
images with a tag the same name as the image repository (ie
`hgmaster:hgmaster`), mirroring the latest result of the
Ansible-on-Docker local build process. We define a bridge
network that all services are connected to, allowing us to
run tests against the cluster from `.t` tests running on the
host. Each machine which will have Kafka installed on it is
assigned a unique `BROKER_ID` which represents it's ID in
the Kafka cluster. Since we want to specify a different broker
ID for each hgweb servie, they are defined as carbon copies
of one another instead of using the `--scale` CLI flag.

An optional environment variable, `MASTER_SSH_PORT` can be
passed to the `docker-compose up` process to allow specifying
the SSH port to be assigned on the host machine. This feature
exists in the current test cluster implementation to allow a
cluster to be started on a port already selected by the Mercurial
test harness as a regex-replacable number. This allows us our
docker containers to print the port in test output and have
the expected result in our test be `$HGPORT`, giving a passing
test.

By passing the variable immediately next to the SSH port
value (22), we make the variable optional as ommitting the variable
will cause the placeholder to interpolate to a blank string,
leaving just `"22"`. This means we can use this docker file
to shutdown a cluster using only the project name (ie without
the master ssh port number) and allows us to test the `docker-compose`
file itself without finding an unused port on the host system.

Differential Revision: https://phabricator.services.mozilla.com/D66164

--HG--
extra : moz-landing-system : lando