testing: add a function to normalize test names for use with `docker-compose` (Bug 1600827) r=mhentges
docker-compose normalizes project names by removing punctuation and
whitespace before assigning them to docker container labels. For
example, a project name like `hello-there!` will be converted to
`hellothere` when used as a label. To allow creating docker test
clusters with unique names in tests, a future commit will teach
the startup script to use the `$TESTNAME` environment variable
as the project name passed to docker-compose. Since we will
want to use the test name to identify which containers belong to
the cluster when orchestrating Kafka startup, we will need to
have a normalized copy of this test name. This commit introduces
a function to strip whitespace and punctuation from the test name,
which will be used by later commits when specifying docker-compose
project names. A docstring test is added to show the behaviour.

Depends on D66160

Differential Revision: https://phabricator.services.mozilla.com/D66161

--HG--
extra : moz-landing-system : lando