ansible/kafka-broker: install Jolokia for Kafka monitoring (Bug 1622136) r=mhentges
This commit installs the Jolokia monitoring agent into the Kafka
process of each Kafka broker. We download the agent from Maven
and place it in the `/opt` directory, next to Kafka and Zookeeper.
We bind the service to port 8778 (the standard port for Jolokia)
so it can be accessed locally by Telegraf for reporting in InfluxDB.
We also add the `LOG_DIR` environment variable to the
`start-kafka` script in the docker test cluster, as this is currently
ommitted and is required for Jolokia.

After this is deployed, the Kafka cluster will need to undergo a
rolling restart.

Depends on D66458

Differential Revision: https://phabricator.services.mozilla.com/D66698

--HG--
extra : moz-landing-system : lando