hgtemplates: update logo image path in `test-static-files.t` r=mhentges
In b285ed119f56 I updated the `header.tmpl` file to use the Mozilla
logo as the OG meta `image` tag, instead of the rather small Mercurial
logo which is intended to be a favicon. We forgot to update the path
in `test-static-files.t` for when the template is rendered, so this
commit does so as a follow up.

Depends on D67021

Differential Revision: https://phabricator.services.mozilla.com/D67022

--HG--
extra : moz-landing-system : lando