hgmo: use `cloud_name` as default key, falling back to `cloud-name` if missing r=zeid
This logic will prevent a failure down the line if we use an
instance data object created from a newer `cloud-init`. At
some point the separator used for the key was changed, and
some versions of cloud-init set both keys. This commit will
make sure we don't break when we hit any of these various
iterations.

Differential Revision: https://phabricator.services.mozilla.com/D68610

--HG--
extra : moz-landing-system : lando