automationrelevance: stop removing `phase` from automationrelevance changesets (Bug 1626641) r=ahal
Before returning data from the `changelistentry` dict in a
response to the `automationrelevance` webcommand, we remove
the `phase` key (among others). Having this key available will
allow us to differentiate between changesets that are part of
the try push and changesets that have been pulled down from
one of the production repositories.

Differential Revision: https://phabricator.services.mozilla.com/D69213

--HG--
extra : moz-landing-system : lando