testing: make `create_virtualenv` support installing the global venv (Bug 1598958) r=mhentges
Make the `name` argument optional and allow passing a `python`
parameter which is intended to be the name of the Python executable
that will be used to create the virtualenv (ie if `python3` is
passed, then the interpreter found on PATH with that name will
be used). We also add the `HGPYTHON3` environment variable for
testing Mercurial 5.0 on Python 3.

Differential Revision: https://phabricator.services.mozilla.com/D69089

--HG--
extra : moz-landing-system : lando