scraper: make `network_border_group` an optional key (Bug 1626722) r=zeid
It seems Amazon has now rolled back the `network_border_group`
key change, which will fail the schema update once deployed.
This commit marks the new key as optional, so we don't break
the scraper when they re-deploy the change. This should be safe
to do since we don't inspect that key anyways.

Differential Revision: https://phabricator.services.mozilla.com/D69597

--HG--
extra : moz-landing-system : lando