robustcheckout: add a `--noupdate` flag to avoid checking out a working directory (Bug 1628019) r=marco
Adds a simple flag `-U` to skip checking out a working directory.
This will be useful for any tool that wants to download a copy
of the repo for analysis, but does not need to operate on a
full working directory (ie BugBug, which needs to analyze patches
and get contents of specific files in memory).

Differential Revision: https://phabricator.services.mozilla.com/D70122

--HG--
extra : moz-landing-system : lando