pash: suggest `Repository Account Requests` Bugzilla component (Bug 1629100) r=jdm
The TinyUrl link that is displayed on hgmo suggests filing
a bug under the `LDAP` component to restore hg access. The
`Repository Account Requests` component is now preferred,
so this commit switches to suggesting that component to
users instead.

Rather than generate another TinyUrl link, we directly
suggest the component to file the bug. Hopefully if a user
needs hg access restored, they know how to file a bug on
Bugzilla.

Differential Revision: https://phabricator.services.mozilla.com/D71187

--HG--
extra : moz-landing-system : lando