url-intercept: patch over bustage caused by chemspill patch (Bug 1541166) r=zeid
During the Python 3 upgrade an issue with the treeclosure hook
was discovered and a minor outage took place while a patch
was created to fix it. In rushing around I patched over an issue
where the `url` was a `Request` object instead of a string as
expected, but did not account for the difference between
`b'url'` and `'url'` on Python 3. This commit wraps the patched
value of `url` to be a byte-string, fixing tests.

Depends on D70957

Differential Revision: https://phabricator.services.mozilla.com/D70958

--HG--
extra : moz-landing-system : lando