testing: update changed test output in newer Mercurials (Bug 1610856) r=zeid
These lines are now printed after pretxnchangegroup hooks have
all successfully run, as shown by the removed lines and different
position of the line when a successful push happens. Likely overlooked
in the same manner as the lack of Python 3 compat in this hook.

Depends on D70962

Differential Revision: https://phabricator.services.mozilla.com/D70963

--HG--
extra : moz-landing-system : lando