vcsreplicator: add a `success` kwarg for use with a repo lock release callback (Bug 1610856) r=zeid
It is possible within the Mercurial internal API to register
a callback to be invoked once a repository lock is released.
Before Mercurial 5.3 this callback is invoked unconditionally.
5.3 added a `success` positional argument to allow certain
callback to not be called if a repo lock cannot be released,
since a failure to release a repo lock likely means some unknown
repo state which should not be acted upon.

This commit adds a `success` kwarg to the `sendheadsmessage`
function, which is passed as a partial function to the repo
callback registrar (`_afterlock`). This will give us compatibility
with both Mercurial 5.2 and 5.3. After upgrading to Mercurial
5.3 we should change `success` from a keyword argument to a
positional one, to ensure the callback isn't run on a repo lock
error.

We mark the vcsreplicator Mercurial extensions as 5.3 compatible
after making this API change.

Depends on D70959

Differential Revision: https://phabricator.services.mozilla.com/D70960

--HG--
extra : moz-landing-system : lando