bundles: update bundlespec to include sparserevlog requirement (Bug 1513276) r=zeid
This commit updates the bundle spec for stream clone bundles
to include the `sparserevlog` repo requirement. After the
repo format upgrade in the upcoming TCW, this requirement
will be present on all stream clone bundles and thus must be
advertised in the clonebundles manifest for each repository.

Since `sparserevlog` was introduced in Mercurial 4.7, the
CDN landing page message is updated to note the requirement
of the newer Mercurial version. The warning for Mercurial 4.1
is removed since most users should be on a newer Mercurial
by now.

I resolved the bundlespec by creating a bundle from a local
repository that already has `sparserevlog` using the exact
bundle command arguments we use in production:

  `hg bundle -a -t none-v2;stream=v2`

and running `hg debugbundle --spec` on the produced bundle.

Differential Revision: https://phabricator.services.mozilla.com/D71342

--HG--
extra : moz-landing-system : lando