configwizard: fix interface bug when curses import fails (Bug 1531059) r=sheehan
This revision fixes an issue with the if statement logic inside `_checkcurses`.
The check worked as expected on Linux machines because `_try_curses_import()`
was returning `True`, and then the if statement passes if the user chooses to enable
curses (`and` statement). On Windows, the if statement immediately fails because
`_try_curses_import()` returns `False` (`and` statement).

With the updated check, the if statement immediately passes if the `_try_curses_import`
fails (i.e. on Windows), and therefore returns without modifying any values. On Linux
the if statement does not pass unless the user answers the prompt with a 'y'.

This check was originally reverting the change that `_checktweakdefaults` was making
in order to set `interface=text` because this check happens after `_checktweakdefaults`
by default.

Differential Revision: https://phabricator.services.mozilla.com/D71404

--HG--
extra : moz-landing-system : lando