testing: create Py2/3 compatible json pretty-print alias (Bug 1541166) r=zeid
On Python 2, `python -m json.tool` prints the json with keys
sorted. On Python 3 the same command prints the json with keys
in the same order as inputted, and has a flag to enable sorting.

This commit adds an alias `ppjson` that papers over the differences
to the tests can be run with either Python version.

Differential Revision: https://phabricator.services.mozilla.com/D74757

--HG--
extra : moz-landing-system : lando