testing: add an `env` command to `./run` (Bug 1598958) r=zeid
To use the Docker test runner, a `.env` file is needed
with several values (gid/uid of the current user, gid
of the Docker group) to allow the container to access
the Docker host and Docker socket from within the container.
This commit adds a `./run env` command that will output
the required contents of this file to stdout, so the file
can be easily created via `./run env > .env`.

To achieve this the test runner is extended to support running
a Python function as a command instead of turning the `./run`
process into a `docker-compose` process via `os.execvp`.

Differential Revision: https://phabricator.services.mozilla.com/D73079

--HG--
extra : moz-landing-system : lando