testing: support passing a more convenient string to `--with-hg` (Bug 1598958) r=zeid
`--with-hg` expects a path to a Mercurial installation which will be
used as the `hg` run in t-tests. Previously when running the tests
from the on-host venv, I would use `fzf`'s fuzzy file finding to
quickly select the version of Mercurial I wanted to run the tests
against. Now that the Docker test environment is the preferred option,
using `fzf` is not feasible.

This commit teaches `run-tests` to accept a Mercurial version as the
`--with-hg` argument. The runner searches in the `mercurials/` directory
of the testing virtualenv to find the executable to be used in the tests.
This allows passing `--with-hg=5.3.2` and makes working with the Docker
test environment much simpler.

Differential Revision: https://phabricator.services.mozilla.com/D73080

--HG--
extra : moz-landing-system : lando