testing: make `run-tests` forget about the various venvs (Bug 1598958) r=zeid
`run-tests` tries to be smart about running tests compatible with
the currently activated virtualenv. This commit removes that checking
code and assumes the `global` virtualenv is being used. This will cause
tests to simply fail or be skipped when run with insufficient requirements,
which should never happen in the Docker test environment.

Differential Revision: https://phabricator.services.mozilla.com/D74760

--HG--
extra : moz-landing-system : lando