testing: update test runner code to be Py3 compatible (Bug 1598958) r=zeid
This changeset includes many small changes to make the test runner
code compatible with Python 3. First we revert the `@check` arguments
to regular unicode strings to be compatible with the earlier `run-test.py`
upstream changes. We add many `decode` and `encode` calls to fix encoding
inconsistencies between Python 2/3. We change the file mode in many
operations from bytes to non-bytes for consistencies with the interfacing
libraries. Calls to a few `dict.keys()` are wrapped in `list` to avoid
issues when indexing the resulting object. All keys passed to LDAP are
converted from byte-strings to regular unicode to conform to that library's
API. Finally, we avoid failing when the Docker state file cannot be read
This should not be a problem since the new test runner looks for an image
tag instead of using the file (all out removal is a pending TODO).

Differential Revision: https://phabricator.services.mozilla.com/D74765

--HG--
extra : moz-landing-system : lando