testing: create native Docker images for test environment (Bug 1598958) r=zeid
This commit converts the test runner Docker image into a Py2 only
image and a Py3 only image. We introduce an `/app` directory and put
the source code and virtualenv in separate directories (`/app/vct` and
`/app/venv` respectively). We set `PYTHONUNBUFFERED=1` instead of
manually modifying `sys.stdout` and remove the `NO_DOCKER=1` environment
variable. We use separate cache volumes for each new image to avoid
package problems in each environment.

After moving the virtualenv out of the `/vct/venv` directory, a few
hard-coded assumptions about the location of the test files need to be
updated. We do so in this commit since it is assumed that running the
tests outside of the Docker environment is no longer supported after this
change.

We also add several new commands to `run` that alias common operations.

Differential Revision: https://phabricator.services.mozilla.com/D74767

--HG--
rename : testing/docker/test-runner/Dockerfile => testing/docker/test-runner-py2/Dockerfile
extra : moz-landing-system : lando