run-tests: add `int_floor` function and make `-j` guessing use it (Bug 1598958) r=zeid
Since I was only running 1 test, I ran `./run tests` without using
`-j` and hit an error since the returned value of `14.0` is not an
integer.

This commit fixes `-j` by introducing an `int_floor` function that
returns the integer floor of a number (ie `14` when `14.5` is passed)
and uses the function on the calculated number of concurrent jobs
to be run when `-j` isn't passed to the test runner.

Differential Revision: https://phabricator.services.mozilla.com/D76369

--HG--
extra : moz-landing-system : lando