pushlog: default to `page = 1` when `node` query string parameter is not a valid int (Bug 1641911) r=zeid
As described in the bug, when the `node` query string parameter
is not a valid `int`, an uncaught exception is thrown. This commit
wraps the `int` conversion in a `try/except` block and defaults to
`1` if the parameter is not an int. This is the behaviour when the
node parameter is omitted.

We also add a test to show the fixed behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D77478

--HG--
extra : moz-landing-system : lando