hghooks: change bug references check to a `pretxnclose` hook (Bug 1607549) r=zeid
After investigating several user reports of the bug references check
hook being too strict and rejecting too many pushes, it was determined
that the `pretxnchangegroup` hooks do not have access to the correct
changeset phase data when executing against pushes to non-publishing
repositories. This means that every changeset in a given push is being
inspected as though it was a draft changeset being pushed to try.
This bug went unnoticed in testing as the test repository is not set to
be non-publishing (as is the configuration for `try`).

This commit begins by setting the test `try` repository to `publishing`,
after which the test fails without any other modifications.

We then create a new `check` type for `pretxnclose` hooks and set them
to only execute when the `txnname` argument is "push". The remainder of
the code for the `pretxnclose` hook is copied from the `pretxnchangegroup`
hook, and the `PreTxnCloseCheck` class is also mostly copied.

We then switch the bug references hook to run as a `pretxnclose` hook, where
it will be able to access the correct changeset phase metadata for inspection.

After applying these changes, the test completes successfully with mostly no
other changes. We add some more verbose output, including the output of `hg out`
to show the phases of the changesets before pushing to the repo.

Differential Revision: https://phabricator.services.mozilla.com/D78726

--HG--
extra : moz-landing-system : lando