hghooks: Don't use `active_scm_*` groups in lando_required hook. (bug 1645168) r=sheehan
By the time a user has reached this hook, we've already verified that they have
an active scm account, so we don't need to check against the `active_scm_*` groups.
Using the `scm_*` groups means that the groups we detect for the user match the
groups that own the hg repositories.

Differential Revision: https://phabricator.services.mozilla.com/D80801

--HG--
extra : moz-landing-system : lando