hghooks: Allow disabling `scm_allow_direct_push` on lando required repositories. (bug 1645168) r=sheehan
For the existing repos where lando is required, there is large group of people
who are allowed to push to the repository, and a small number of people that,
due to the nature of their work and frequency of their work, need to be able to
push to the repository directly. They have been added to the `scm_allow_direct_push`
group, which bypass the requirement to provide a justification for bypassing lando.

For the `ci-admin` and ci-configuration`, there will be a small number of
people with push access, and none of them will need to regularlly push outside
of Lando. This adds a new configuration, that allows the lando required check
to not treat members of `scm_allow_direct_push` specially.

Differential Revision: https://phabricator.services.mozilla.com/D79844

--HG--
extra : moz-landing-system : lando