testing: add `active_scm_autoland` to testing LDAP config (Bug 1653321) r=tomprince,zeid
In changeset f6ab42087b836792 we changed the matching string for
adding a user from `scm_level` to `scm_`. The previous behaviour
meant we did not add users to `active_scm_autoland` when adding
to `scm_autoland`, and as a result the `active` level did not exist
in the testing LDAP config.

This commit adds `active_scm_autoland` to the testing LDAP config.

Differential Revision: https://phabricator.services.mozilla.com/D83811

--HG--
extra : moz-landing-system : lando