configwizard: suggest `mach install-moz-phab` in codereview step (Bug 1663799) r=sylvestre
The Mercurial config wizard has a small paragraph of help text pointing users
to the Phabricator user guide when the user indicates they will be submitting
changesets to Firefox. On top of doing this, we should suggest users run
`mach install-moz-phab`, to improve discoverability of the command and
to avoid requiring the user read the entire guide to find this useful
tool.

Differential Revision: https://phabricator.services.mozilla.com/D89612

--HG--
extra : moz-landing-system : lando