firefoxtree: change `outgoing.missingheads` to `outgoing.ancestorsof` r=zeid
This field was renamed in Mercurial 5.5. A deprecation warning indicates
the field never contained what it described and code that uses it should
be checked for correctness. Despite the incorrect name, the associated
test (`test-prevent-push-multiple-heads.t`) fails when removed. Renaming
to the new name keeps the same behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D88681

--HG--
extra : moz-landing-system : lando