hgwsgi: remove dead entries from `hgwsgi/releases/hgweb.config` (Bug 1667731) r=zeid
Removes some entries for repositories that no longer exist on hgmo.
The existance of these entries is causing an exception in core
Mercurial code and causes https://hg.mozilla.org/releases to fail
to render.

The upstream issue will need to be addressed too, but this is a quick
fix that we should address in the meantime.

Differential Revision: https://phabricator.services.mozilla.com/D91607

--HG--
extra : moz-landing-system : lando