ansible: use stream clones in project repo reset playbook (Bug 1672877) r=zeid
Currently when we reset a twig (project branch) on hgmo by running the
Ansible playbook `hgmo-reset-twig.yml`, the repo is reset by creating
a new empty repo on hgssh and locally pulling from `mozilla-central`.
This creates an `hg-repo-init-2` message in vcsreplicator so the repo
is created on hgweb, after which we run `hg replicatesync` to trigger
a full sync of the repo. This full sync causes all hgweb nodes to pull
from hgssh to gather the new data.

This process is slow, and causes the vcsreplicator queue for the project
repos to lag far behind. This triggers many alerts in #vcs on Slack,
and can cause issues for others working on project branches.

This commit updates the playbook to instead create a fresh clone of the
repo using stream clones on both hgssh and hgweb, speeding up the process
and working around the vcsreplicator queue lag. As a result of this change,
we also run the `repo-permissions` script on hgweb and earlier in the reset
process on hgssh. I used this script to resolve bug 1672465 without
running into the aforementioned problems.

Differential Revision: https://phabricator.services.mozilla.com/D94523

--HG--
extra : moz-landing-system : lando