vcsreplicator: teach consumer processes to handle abandoned transactions (Bug 1679880) r=mhentges
Implements autorecovery from abandoned transactions in repos on
hgweb mirrors. We create a decorator for the main vcsreplicator
message handler that catches a Mercurial command error and looks
for the string `run 'hg recover'`, running the aforementioned
command if the string is present and re-processing the message.
If `hg recover` fails or the second attempt to re-process the
message fails, the replication process will error out.

Differential Revision: https://phabricator.services.mozilla.com/D98238

--HG--
extra : moz-landing-system : lando