ansible: set apache `MaxClients` to 25 on private hgweb mirrors (Bug 1481291) r=mhentges
The current calculation we use sets the `MaxClients` value to just 18 on
the private mirrors. Looking at the logs in Chronograf I can see that the
occurrances of 500s generally happen when we reach the max number of clients.
When we hit the maximum the instances still have CPU and memory to spare,
leading me to believe the Apache config should be tweaked to allow more
concurrent requests.

Also remove an unnecessary comment while we're here.

Differential Revision: https://phabricator.services.mozilla.com/D98981

--HG--
extra : moz-landing-system : lando