terraform: add a new hgweb backup node (Bug 1689640) r=mhentges
Adds a new variable `backup_node` to the `aws-vpc` module which creates a
backup node in the VPC when set. This backup node is the same as the
other hgweb mirrors in the VPC (has access to MDC1 via the corporate
VPN) except it has much more disk space and is not connected to the
load balancer through which CI sends it's VCS queries. This will allow
the node to replicate all changes made to repos in MDC1 and have the
replicated repos be restorable via bundling/cloning from the backup
node.

Sets the new `backup_node` value to `1` for the us-west-2 instantiation
of the `aws-vpc` module.

Depends on D103912

Differential Revision: https://phabricator.services.mozilla.com/D103913

--HG--
extra : moz-landing-system : lando