vcsreplicator: avoid processing `hg-repo-delete-1` message if vcsrepicator "backup" config set (Bug 1689640) r=mhentges
Adds an `is_backup` function to vcsreplicator's `config` object
which checks the config for the presence of the `consumer.backup`
option. If this function returns `True` when processing a repo
delete message, it is ignored on the node.

Add this config option to the backup nodes so repos aren't accidentally
deleted.

Depends on D104834

Differential Revision: https://phabricator.services.mozilla.com/D104835

--HG--
extra : moz-landing-system : lando