vcsreplicator: add a short summary of message payload to logs (Bug 1695029) r=mhentges
vcsreplicator queue logs are currently useful for knowing what is
causing a queue to jam. As it stands now they are only semi-useful
at establishing timelines of when certain messages are processed
in the queue, since the logs only indicate the type of message being
processed instead of data within the message.

This commit adds a mapping of message type to a format string, and
a function which splats the contents of the JSON payload into the
format string, for listing in the vcsreplicator logs.

Differential Revision: https://phabricator.services.mozilla.com/D107697

--HG--
extra : moz-landing-system : lando