pushlog: add a `branch` query string parameter to filter HTML pushlog (Bug 1711742) r=mhentges
Adds capabilities to filter returned pushlog results by a given Mercurial branch.
Checks for a `branch` query string parameter in the pushlog feed extension which
is passed through to the `repo.pushlog.pushes` query function. If the value of
this parameter is not `None`, we assert the value is a valid branch and filter
returned pushlog results that exist on the given branch.

Differential Revision: https://phabricator.services.mozilla.com/D115490

--HG--
extra : moz-landing-system : lando