testing: install Mercurials by cloning `hg-committed` repository (Bug 1727330) r=zeid
The `hg` repo is supposed to have all the work included in `hg-committed`,
but in reality `hg-committed` is the source of truth and is kept more up
to date. At the moment `hg` is missing the `5.9rc1` release tag, so switch
to using it to install Mercurial.

Differential Revision: https://phabricator.services.mozilla.com/D123467

--HG--
extra : moz-landing-system : lando