hgmolib: open file in regular mode when parsing `requires` files (Bug 1727342) r=zeid
The value for `--requirement` was being passed as a `str`, while
the was being read and each line in the file was passed as a `bytes`,
leading to comparisons like `"store" == b"store"`, which is `False`.
Opening the file in regular read mode makes both values `str` and the
comparison is correct again.

Depends on D123479

Differential Revision: https://phabricator.services.mozilla.com/D123480

--HG--
extra : moz-landing-system : lando