vcsreplicator: use `phaseroots` integer keys directly instead of implicit enumeration (Bug 1727344) r=zeid
Mercurial internals formerly defined phases via an expansion of `range(3)`,
mapping the phases to each integer. Internally the integers are used so
the phases have an order (ie `draft < public`), but these phases are mapped
to names that we use in the Mercurial command line. At some point new phases
were added with their own unique values that break the `range(3)` standard.

In `vcsreplicator` we would read the `phaseroots` dict and enumerate the keys
to determine which phases are present on a given repo, however the presence
of the new phases breaks this assumption. Since the `phaseroots` object is
aleady a mapping of the phase integer representation to revision hash, we
should use the keys of the `phaseroots` mapping directly to determine which
phases are found on a given repo.

Depends on D123481

Differential Revision: https://phabricator.services.mozilla.com/D123482

--HG--
extra : moz-landing-system : lando